feat: :bug: Accumulated ratio accounts for position weights, other parts of algorithm now work with multiple positions
Fixed calculateOptimalRatio to simply add up the tokens that will be used by each position, then convert that into a ratio at the end, in order to preserve a ratio which accounts correctly for the weights of different liquidity positions. Fixed rest of alpha-router to correctly handle multiple liquidity positions. Fixed some tests with strange behaviors when the ratio is (0, 0).